only releases have stages
