+++ ruby/installations/release.rb
@@ -8,7 +8,7 @@ module Installations
     def collaborator_map
       @installation_collaborator_map ||= super.merge(
         mandatory_keys.reduce({}) do |m, k|
-          m[k] = protoype_for(k)
+          m[k] = collaborator_for(k)
           m
         end.compact
       )
