+++ ruby/releases/collaboration.rb
@@ -5,37 +5,33 @@ module Releases
 
     relation_accessor :predecessor
     relation_accessor :assembly
-    relation_accessor :stages
 
     delegate([:outline, :collaborating_divisions] => :schema)
 
     def memento; OpenStruct.new(to_h) ;end
 
-    def collaborators
-      @collaborators ||= [stages&.map(&:collaborators), collaborator_map.values].flatten.compact
-    end
-
-    def stages
-      @stages ||= struct.stages&.map { |s| stage_for(s) }
-    end
+    def collaborators; collaborator_map.values ;end
 
     def collaborator_map
       @collaborator_map ||= keys.inject({}) do |m, k|
-        m[k] = protoype_for(k)
+        m[k] = collaborator_for(k)
         m
       end.compact
+    end
+
+    def collaborator_for(key)
+      collaborating_divisions[key]&.prototype(collaboration: self, label: key)
     end
 
     def to_h
       keys.inject({}) do |m, k|
-        m[k] =
-        if k == :stages
-          stages.map(&:memento)
-        else
-          collaborator_map[k]&.memento || struct[k]
-        end
+        m[k] = memento_for(k)
         m
       end.compact
+    end
+
+    def memento_for(key)
+      collaborator_map[key]&.memento || struct[key]
     end
 
     def schema_keys; schema.keys ;end
@@ -47,21 +43,6 @@ module Releases
       else
         assembly&.send(m, *args, &block) || super
       end
-    end
-
-    protected
-
-    def stage_for(struct)
-      stage_class.new(struct: struct, assembly: self)
-    end
-
-    require_relative 'stage'
-    def stage_class; Stage ;end
-
-    private
-
-    def protoype_for(key)
-      collaborating_divisions[key]&.prototype(collaboration: self, label: key)
     end
 
   end
